Created new module air_quality.py
